Add support for scanning multiple directories
Update the main routine to consider `argv` as a list of projects to scan
and scan each one provided (still defaults to the current working dir if
no argument is provided). Also, if multiple dirs are scanned, each one
will be listed in stdout in order to separate findings clearly. Lastly,
if an unexpected error occurred for any of the dirs the remaining dirs
will still be scanned before exiting with an error.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>