Scan Action manifest at the root of a project
Extend the functionality of the scanner to look for an Action manifest
in the root of the repository and, if it exists, parse it (using a
specialized parsing function) and process it (using a specialized
processing function). The new functionality is tests in line with
existing test practices.

Additionally, there's one typo fixed and one improvement suggested by
gocritic (https://go-critic.com/), both in `workflow_test.go`, as well
as the removal of a redundant newline in `workflow.go`.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>