Scan a specific file
Extend the functionality of the scanner to support scanning a specific
file. This is achieved by first checking if a given path is a file and,
if it is, processing it directly as either a manifest or workflow file.
If the path was a file, it will not be considered for `%s/action.yml` or
`%s/.github/workflows/*.yml` (and if there's additional paths, it will
continue scanning those).

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>