Refactor `main.go:run`
Update the `run` function in `main.go` to reduce its cognitive
complexity as well reduce code duplication (in terms of printing
problems as well as parsing and processing files). Additionally, the
argument name has been updated from the old and outdated `wd` (short for
"working directory") to "target" (aligning it with how it's called).

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>