Don't exit early if an Action manifest cannot be parsed
Update the behavior of the scanner to not stop analyzing a working
directory early if an Action manifest failed to parse. This aligns the
behavior for Action manifests with that of workflows (the scanner
continues if any one workflow fails to parse). This should also improve
reports by being able to find dangerous uses of expressions in workflows
without having to make sure the Action manifest is valid (mainly useful
for analysts).

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>