Exit with a non-zero exit code when problems are detected
Update the `main` function to track whether or not problems were found
for any of the targets and exit with a (new) non-zero exit code if that
is the case.

Also, make it so that an error for any target results in a non-zero exit
code. Before this change a non-zero exit code would only be used if the
last target resulted in an error.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>