Test CLI with `testscript`
Add an initial set of `testscript`-based tests to test the logic of the
`main.go` file. This testing utilities provides a DSL for writing test
for CLI applications. It does not quite run the project as a compiled
CLI, instead through the addition of the new `ades` function it can run
what is effectively the CLI (barring `main`) while keeping the ability
track and report code coverage.

These tests also uncovered a problem in the implementation where trying
to process a file as a workflow following a problematic manifest would
reset the `hasProblems` and thus not exit with a non-zero exit code.

To find out more about `testscript` read either the official docs:
<https://pkg.go.dev/github.com/rogpeppe/go-internal/testscript>, or
these blog posts: <https://bitfieldconsulting.com/golang/test-scripts>.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>