Export & document parsing logic
Update `parse.go` to intentionally export parsing logic in order to
allow re-use of the logic for parsing GitHub Actions workflow and
manifest files. Even if incomplete, it may be re-usable, plus I'm
willing to expand the structs to allow more use cases (it may come
in handy for this project in the future). As exported funcs (and
structs) there have been given documentation for completeness.

The rest of the codebase is kept unexported. While the parsing logic
is unlikely to see a breaking change (with the struct renames), the
rest of the codebase is, at this point in time, still unstable and
likely to have API changes.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>