Align `tryManifest` and `tryWorkflow` behavior
Update the implementation of `tryManifest` to behave similarly to
`tryWorkflow` (and, in my opinion, more intuitively). Now, both `func`s
will error if there's no file to read at the provided path.

To keep the CLI behavior the same, `analyzeRepository` func is updated
to only print an error if the observed error *isn't* a reading error
(because in the context of analyzing a repository, we don't care if the
manifest doesn't exist).

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>