Align documented license with license header, add missing license header
Per the license headers in the `*.go` files (one of which was missing),
the license used by this project is `GPL-3.0-or-later` (as opposed to
`GPL-3.0-only`) per the inclusion of "or (at your option) any later
version." line in the first paragraph.

Note that this is not a license change, but rather a documentation
correction. The license file (COPYING.txt for this project) and per-file
license headers are authoritative w.r.t. the README.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>