Add suggested fix messaging to tool output
Update the tool so that it outputs suggested fixes for violations it
finds. This should help users to mitigate any potentially problematic
expressions in their workflows or actions manifests. The suggestions
seen here are somewhat crude, but should get the job done.

This implementation is based on an adjustment to the analysis code
and violation struct, updating it to record the "kind" of the violation.
This "kind" value is than used by the violation reporting logic to
provide a suggestion specific to the "kind" of violation. This kind
value may turn out to be useful for other context-dependent feature as
well.

Also adjust the tool input to use indentation that is a multiple of 2
instead of 3 to reduce line length for the newly introduced double
indented text.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>