Implement machine readable output in the form of JSON output
Add a new flag to the CLI named `-json` which can be used to change the
output format to a JSON string that can be used by machines to use the
results however they like. The format is explicitly defined in the
`schema.json` file.

A basic test for the `-json` functionality is included mainly to ensure
that, when used, only JSON content is outputted by the program. On top
of that, the `make test` command also verifies the JSON schema defition
is always valid using the `jv` CLI tool [1].

Besides being easier for machines to read, the intention behind the JSON
output is also to be more stable. I.e., the intent is that breaking
changes are introduced to the JSON schema less often, which should make
it easier to integrate with `ades`. That being said, the schema has been
written to explicitly allow `additionalProperties` on any object so as
to allow for extension of those objects without breaking the schema.

--
1. https://github.com/santhosh-tekuri/jsonschema

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>