Set up mutation testing with Ooze
Add a new make target for running mutation tests [1] with Ooze [2].

The goal of using mutation testing is to improve the quality of the test
suite. This is achieved by determining if small changes in the source
code (called mutants) are detected by the test suite. If they're not it
is an indication that the test suite does not assert anything accurately
or meaningful about that line of code (even though it might be covered).

Mutation tests are also run in continuous integration to ensure quality
of the test suite over time. A zero-mutant policy will be adopted in
order to simplify usage and maintenance of the setup. Note that as of
this commit the project still has 10 mutants that need to be eliminated
through better testing (or refactoring).

--
1. https://en.wikipedia.org/wiki/Mutation_testing
2. https://github.com/gtramontina/ooze

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>