Separete aggregate violations by target and include target in JSON data
Update how JSON data is reported to include the target to which the
problem/violation belongs. This change has two motiviations, both
concerning the use case where more than one target is scanned at once.
First, it avoids overriding results when two targets contain the same
file. Second, it makes it possible to determine exactly which file a
problem is for (without the target you may not know which project to
look at).

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>