Treat a workflow named `action.ya?ml` consistently
Update the scanner implementation such that a workflow named
`action.ya?ml` (i.e. it is in the `.github/workflows` dir) is always
analyzed as a workflow. This is achieved by considering the full path of
the file when determining whether to scan a given file as a workflow or
as a manifest.

The added test case exists solely to ensure this change holds.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>