Set up mutation testing with go-mutesting
Update the make target for running mutation tests to run go-mutesting[1]
instead of Ooze[2]. The motivation for this is to check out the
difference between the two mutation engines and switch if the new engine
is "better" than the old one.

The zero-mutant policy will be maintained in order to simplify usage and
maintenance of the setup. Note that as of this commit the project still
has 21 mutants as identified by the new mutation engine. These will need
to be eliminated through better testing (or refactoring).

--
1. https://github.com/zimmski/go-mutesting
2. https://github.com/gtramontina/ooze

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>