Refactor source code and improve tests
Source code wise, `analyze.go` is refactored to simultaniously simplify
`analyzeStep` as well as avoid "unnecessary" statements due to default
values for `kind` and `script`.

Test wise, the unit tests for `analyzeJob` have been updated to assert
something about the opening `if`-statement which was previously covered
but not checked. Related to this change, all unit tests now separately
define a `TestCase` type (if used) to avoid having to extract it later.
Separately, integration tests are updated to 1) cover the human-readable
output for GitHub script violations and 2) ensure that directories with
a `.yml` or `.yaml` "extension" are ignored.

These changes are cherry-picked from the work integrating `go-mutesting`
into the project. It addresses all mutations found in the current source
code as identified by `go-mutesting` in code that is covered by tests.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>