Initial support for analyzing from stdin
Add (and test) support for analyzing a file from stdin. Because we can't
know whether the data from stdin is a manifest or workflow, it will be
analyzed as both and report violations if either interpretation is
problematic.

The behavior is such that if and only if the `-stdin` flag is used, the
program will read from stdin. This was chosen because dynamically
detecting if there is data on stdin was found to be unreliable. Hence,
for now the user needs to be explicit about their intent. If a way is
found to reliably detect data on stdin, the implementation can be
updated to use that.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>