Define containerized version of the CLI
Add a Containerfile (engine-agnostic Dockerfile) that defines how to
build a containerized version of the ades CLI to broaden support (e.g.
to those that would like to build from source but don't have the Go
toolchain installed). The Makefile now defines a `make container` target
that performs the conanical container build for the project. The README
also states that using the container is an option.

The container build as well as it's functioning is continuously tested
in the "Check" GitHub actions job. This job depends on the build working
and tries to build the container using `make container` and run it using
the `--help` argument. This covers both Docker and Podman in order to
try and be engine agnostic.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>