Continuously validate JSON data against the schema
Add a test that validates arbitrary instances of `jsonOutput` against
the schema definition in `schema.json`. This is aimed towards avoiding
a situation where the actual JSON outputted by the CLI doesn't match the
stated schema.

In this process, the JSON schema itself is also validated. Hence, the
separate validation done as part of the `make test` target has been
removed (and correspondingly the import of `jsconschema/cmd/jv` in
`tools.go` as well).

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>