Make suggestions optional
Remove the by-default-outputted suggestions in favor of optional
suggestions that can be enabled with the new `--suggestions` flag. When
not enabled, the output will instead contain (new) rule IDs, which are
(now) present in the documentation and can be used to look up what to
do for a given error (just like suggestions).

The main motivation for this is to make the output less noisy by
default. In particular, when one already knows how to fix a given error
the suggestion is just noise and makes it harder to parse the output.
Additionally, the reference to the documentation allows users to cross
reference the violation with a more detailed explanation (including
example fix/suggestion) from the documentation.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>