Run table driven tests in parallel
Update all table driven tests to use `t.Parallel` in order to run them
in parallel (when possible). Also start using the paralleltest tool
(<https://github.com/kunwardeep/paralleltest>) to detect incorrect use
of `t.Parrallel` (not currently used to enforce all tests to be run in
parrallel, this can be re-evaluated in the future).

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>