Add `test-randomized` command
Add a new helper command to run tests in a randomized order. This is
helpful for guarenteeing tests are independent, which is an important
property for correctness and reproducibility.

The command is used in CI in order to try and detect flakiness due to
inter-test dependence. It's not the default test command for local
development because it lacks test results caching, which is important
for development speed.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>