Add and configure Capslock
Add Capslock (<https://github.com/google/capslock>) as a tool dependency
and use it to generate a capabilities list and track it over time, in
particular flagging when capabilities change. The idea behind doing this
is to ensure all potentially dangerous function being used are known and
make sense in the context in which they're used.

The capabilities.json file is included in the commit history because it
is needed to do a comparison from one version to the next.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>