Use `google/go-licenses` for license compliance
Add `google/go-licenses` as a tool dependency to support performing
continuous and automatic license compliance. It is used through a new
Make command called `compliance` (run as part of the general `verify`
command) that checks runtime dependencies (i.e. those that will be part
of the compiled binary) use allowed licenses. The compliance is enforced
continuously through a new continuous integration job.

The current allowed license list is the minimum required. This list may
be expanded upon the addition of a new dependency if it can be complied
with.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>