Disable verbose logging when running mutation tests
Update the `test-mutation` command to no longer use verbose logging.
This follows the continuous integration reporting on the errors produced
as part of the logs due to (eliminated) mutants, which make the CI less
useful. In practice, I've not found the verbose logging particularly
useful (apart from the fact that it indicates progress).

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>