Detect use of expression on known vulnerable inputs
Expand the scope of this scanner by also looking for the use of action
expressions in Action inputs that are known to be vulnerable to
injection, starting with `ericcornelissen/git-tag-annotation-action` as
a proof of concept.

The difference between these violations and those previously covered is
that it only concerns Action inputs which shouldn't be dangerous, but
are in practice because of a bug (vulnerability) in the action. Whereas
using expressions in `run:` and `actions/github-script` is always
dangerous.

The current implementation is limited to only the `uses:` syntax with
full versions, not SHA pins or moving major version refs.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>