Refactor to allow for easier extension
Change the implementation of `analyze.go` with the aim of making it
easier ot extend. In particular the goal was to make it easier to add a
new analysis to the `analyzeStep` function, all other changes are in
support of that. Of particular note is the deduplication of the loop
over violations (to inject metadata) as well as the omission of helper
functions to determine the step type (mostly).

On the same note, `parse.go` is refactored as well. First the `StepWith`
struct was removed in favor of the inherently extensible `map` type.
Second, it now provides a function to parse the 'uses:' value of a job
step into the action name and its version - which is already used
internally.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>