Make tooling exit with an error code
As much as supported, configure tools to exit with a non-zero exit code
in order to ensure problems are seen and fixed, as well as cause CI to
fail.

Most tools do this by default, the changes here instruct tooling that
have optional support for it to do it. Unfortunately this still leaves
some tools that never exit with a non-zero exit code (e.g. `godot`).

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>