Use betteralign for improved memory usage
Add betteralign as a tool dependency and use it when vetting to find
structs that aren't declared in a way that optimizes memory. It can
automatically fix such structs by re-aligning the fields using the
`-apply` flag.

The tool found two structs with fields that aren't aligned for optimal
memory usage, these have both been updated (automatically).

This also updates nilaway and its transitive dependencies because
otherwise nilaway would exit with an error due to an incompatibility
error.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>