Change process from stdin flag from `-stdin` to `-`
Replace the `-stdin` flag by `-` for reading a file from stdin. This is
inspired by the same strategy used by some other CLI tools (whereas I've
never seen an `-stdin` flag for this purpose).

Because of the way Go's `flag` package this can't be implemented as a
flag in that sense. Hence, this uses a custom implementation that looks
if the provided arguments is a literal "-" character. The current
implementation intentionally only considers the first argument for this
purpose and ignores any other provided tags if it's there (though I may
be convinced to change that).

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>