Use `whitespace` for consistent use of newlines
Extend list of tools used for vetting to include a tool called
whitespace (<https://github.com/ultraware/whitespace>), it "is a linter
that checks for unnecessary newlines at the start and end of functions,
if, for, etc.". This code base currently has no violations w.r.t. this
tool so no further changes are included.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>