Create `make release` to aid in the release process
Add a new make target called `release` which initiates the release
process by carrying out steps 1-4 automatically and providing directions
on how to proceed after that. The goal is to (incrementally) make
releases easier to do and hard to mess up.

To support this automation the implementation is refactored slightly to
create a clear anchor in the codebase that the automation can use to
perform a replacement of the version string.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>