Upgrade to Go 1.22.0
Upgrade from Go 1.21.5 to 1.22.0 to stay up-to-date with the Go
language.

One important change in this upgrade is the behavior of loop variables,
which are now recreated every loop, removing the need for binding it
within the loop as `x := x`. Hence, this commit includes refactoring to
remove such statements and also removes the looppointer static analyzer
which exclusively scanned for this problem. The paralleltest static
analyzer is temporarily disabled because it includes a check for that
which cannot be disabled - the intent is to update this tool upon an
update that makes it compatible with Go 1.22.0.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>