Migrating `Makefile` to a Go-native taskfile
Remove the `Makefile` and introduce `tasks.go`. A task-runner written in
Go. It's a lightweight task runner that only requires Go, thus removing
this project's dependency on GNU Make. I implemented it for this express
purpose.

Is it really necessary to remove the dependency on `make`? Perhaps not,
but I think it's an interesting approach to reducing the project's
prerequisites, and I like having as few prerequisites as possible. Also
the challange of building this was a fune one :)

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>