Audit only when necessary
Update the `audit.yml` workflow to only trigger for Pull Requests when
there is a chance for a new finding as a result of the change. It is
only possible to introduce a new vulnerability when the dependencies
change (covered by `go.mod` and `go.sum`) or when the code changes due
to usage of a new third-party API (covered by `*.go`).

The motivation for this is to reduce unexpected CI failures. For example
when only a text/MarkDown file is changed the result of this workflow is
not interesting.

We also always run this workflow when it is changed as a sanity check.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>