Add rules for roots/issue-closer
Extend the rules of ades to include two dangerous uses of workflow
expression in the roots/issue-closer action [1]. The implementations are
based on the guidelines for new rules from the Contributing Guidelines.
These rules were added based on their inclusion in [2] after validating
the impact (and suggestion) with a workflow containing something like:

    - uses: roots/issue-closer@v1.1
      env:
        EXAMPLE: foobar
      with:
        issue-close-message: |
          Test from expression '${{ github.event.issue.title }}'.
          Test from eval '${eval('"x"')}'.
          Test from environment variable '${process.env.EXAMPLE}'.

Beyond adding the rules this commit includes two further contributions.

First, it updates the guidelines for adding a new rule in two ways. It
corrects the name of the `Violation` type following the changes of [3]
and it also adds the requirement that the new rules must be added to the
`RULES.md` documentation file.

Secondly, the `rules_test.go` file now delivers on the promise that "The
rule id, title, and description are tested automatically." through the
`TestAllRules` test suite that asserts properties that should hold for
all rules.

1. https://github.com/roots/issue-closer-action
2. https://boostsecurityio.github.io/lotp/
3. aea637182815d57705e13c217dc00d6d610c2334

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>