Find manifests anywhere in a repository
Update the way repositories are handled to walk the repository file tree
to look for Action manifests. In particular, it will look for an action
manifests everywhere except in the `.git` directory.

Relatedly, the way the file system is walked for workflows is updated to
match the implementation of that for manifests.

Lastly, this replaces the use of the standard library `path` by
`filepath` for better compatibility and more consistency.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>