Use custom ids to avoid the map of rule IDs to links
Simplify the website code and reduce the work needed to add a new rule
by leveraging custom ids in the `RULES.md` documentation to create
deterministic anchors.

In particular, the `id` attribute on the added `<a>` tags create an
anchor on the generated HTML that can be referred to (whereas the
default anchor for a title is a hyphenated version of the name). Thus,
this allows linking to the custom anchor which is easier to determine.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>