Harden the website against XSS
Improve the website to properly construct dynamic content in a way that
avoids web attacks - even if they (currently) have no impact on this
website.

An XSS proof of concept for the website prior to this change would be
a workflow like:

```yml
name: XSS Example
on: [push]

jobs:
  example:
    runs-on: ubuntu-latest
    steps:
    - name: Unsafe run
      run: echo 'Hello ${{ inputs.name.<b>xss</b> }}'
```

which will render the text "xss" in bold. This will now be rendered
as exactly "<b>xss</b>".

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>