Set up container signing with Cosign from GitHub Actions
Update the "Publish / Docker Hub" job to sign container releases using
Cosign (<https://docs.sigstore.dev/signing/quickstart/>). In particular,
this uses keyless signing based on the OIDC token available in the job.
That way, the published container is linked to the workflow that created
it.

The version of Cosign to be used is hardcoded in the workflow. This is
helpful w.r.t. to reproducibility (i.e. we know exactly what version
will be used, assuming `sigstore/cosign-installer` is benign) but limits
the ability to stay up-to-date because (as far as I know) there is no
existing way of recieving automated update for this value.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>