Add `-conservative` option
Update the CLI with a new `-conservative` option that allows users to
only get warnings about GitHub Workflow Expressions that are known to
be controllable by attackers and therefor higher risk.

This option should be useful for users that want to know only about
high risk expressions as well as researchers that want to run this tool
on many projects and get less overwhelmed.

This list of expressions considered in conservative mode covers all of
the expressions considered by:
- ARGUS: A Framework for Staged Static Taint Analysis of GitHub
  Workflows and Actions
- CycodeLabs/raven "Injectable X" queries.
- Semgrep `run-shell-injection` rule for GitHub Actions.
- Actionlint "Script injection by potentially untrusted inputs" check.
- `githubuniverseworkshops/workflow-script-injection` exercise 1.
- `binarytrails/notes` notes on GitHub Actions.
- checkov rule at `github_actions/common/shell_injection_list.py`.
Except for `github.event.inputs ...` which is only included by Semgrep.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>