Change wording of "Workflow Expressions" to "GitHub Actions Expressions"
Update wording related to the term "Workflow Expressions" throughout the
project to use "GitHub Actions Expressions" or just "Expressions". This
is motivated by 1) the fact that the page linked for this term calls it
just "Expressions" and 2) expression can be used in action manifest in
addition to workflows. The "GitHub Actions" prefix is used in certain
cases to be more clear about what kinds of expressions are considered.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>