Implement automatic fixes for ADES102
Extend the functionality of the CLI with an (experimental) autofix
feature and implement an autofix mechanism for ADES102. To achieve
this the CLI is extended with a new flag; the rules logic is extended
with a (optional) fix function per rule; ; the violation logic to keep
track of the `source` where the violation occured, as well as the unique
job and index identifier; and the parse logic to include the `env:` map
for steps.

The feature is marked as experimental in both the docs and the CLI. In
particular the CLI flag is marked as experimental to make sure anyone
using it is aware it's experimental, which is important in this case
because it may change files on disk (possibly incorrectly).

The fix logic is based on a search and replace strategy leveraging
regular expressions to 1) handle arbitrary whitespace and repetition in
the source YAML, and 2) match indentation of the existing YAML. This
approach was chosen over modifying the struct marshalling to YAML
because that is very likely to mess up the formatting w.r.t. the
original file.

The changes to parsing and creating violations support the generation of
fixes. The parsed env map is used to extend add new environment
variables if necessary. The jobKey and stepIndex are computer-usable
unique identifiers for the job and index in which the violation occurs
(whereas JobId and StepId are human-readable but not guaranteed to be
unique).

All changes are extensively tested. Both unit and testscript tests cover
the new and changed functionality. Especially the testscript tests make
sure the regular expressions used for replacement work for valid
manifest and workflow files.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>