Add related work `poutine`
Add a recently released related tool called `poutine` from
BoostSecurity.io to the related work section of the README.md as it,
among other things, looks for problematic expressions in GitHub Actions
workflow.

The text was taken from `CycodeLabs/raven`, and its text has been
adjusted to be more nuanced as a result of and to be in line with the
text for `poutine`.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>