Support major version-only refs when checking if a rule applies
Update the implementation of rules to support comparison against refs
consisting only of the major version (e.g. `v4` in `foo/bar@v4`). In
particular, the implementation assumes the major version always points
to the latest release for said major release, thus if the major version
number equals or exceeds to vulnerable version the use is of the action
is safe, otherwise it is not.

Crucially, when the major version equals the vulnerable major version
this works because of the following: the internals compare against the
*patched* version of the action. If the patched version is in the same
major version range, allowing use of that major version is OK. If the
first patched version is the next major version, than only the next
version will be considered unaffected.

This change is tested with one modified and two new test cases ("Major
version only, same major version", "Major version only, earlier major
version" & "Major version only, later major version" resp.). On top of
that, this improves the test suite for the relevant function with more
unique test cases of different scenarios when the full semver string is
used, in particular when it comes to which of the MAJOR.MINOR.PATCH
numbers differs between the `Ref` and comparison version.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>