Add support for comparing to major+minor semver notation
Improve the internal logic of rules that look at the version of an
Action with support for major+minor semver notation. That is, the
version used is specified as vX.Y (instead of vX or vX.Y.Z). I have
seen this supported by some Actions which is why I'm adding support
for it.

The way this work is mostly straigtforward. All earlier major versions
as well as same major version with earlier minor version are considered
"before" the target version. All later major versions as well as same
major version with later minor versions are considered "not before" the
target version. In case the major version and minor version match,
however, the version is considered "not before" the target version
because the version range defined by the use of major+minor notation is
the latest patch version for vX.Y, which should be unaffected if the
major and minor versions are the same between the query and target
version.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>