Configure user of the ades Container to be non-root
Update the `Containerfile` to create and use a non-root user as a
defense-in-depth strategy. This is based on a Docker scout report for
this image on Docker hub. The solution is based on the provided
remediation link:
<https://docs.docker.com/scout/policy/#default-non-root-user>
and was created with the help of ChatGPT(-4o) (used to find potential
implementations of using a non-root user in a `FROM scratch` image and
find ways of verifying the changes too).

I tried to verify the configuration from within the container but mostly
failed (when changing `FROM scratch` to alpine and starting with
`/bin/sh` the current setup gave a non-root user in a root group; in
`FROM scratch` I was unable to verify it directly and failed to do so
with Go's `os/user` package; in the end I used `docker inspect
--format='{{.Config.User}}' to verify it externally).

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>