Automate `-version` value
Update the build procedure for ades to dynamically determine the version
based on the VCS information available. The logic is specified in
`TaskVersion` but broadly speaking tries to get the version based on a
tag (or else related commit information) and if the project is dirty. As
a result, the behavior is unchanged if building from a clean commit with
a tag, else the commit info and/or dirty state information can help with
diagnosing issue (if ades wasn't build from a release tag or with a
dirty state, that's a good starting point).

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>