Add support to the `isBeforeVersion` function for SHAs with annotation
Improve the rules that depend on the internal `isBeforeVersion` function
by adding support for SHA-based refs with comments/annotations to it. It
will now, if the ref isn't parsable as a semver string, try to use the
comment/annotation (i.e. the comment on the line of the `uses` directive
after the value, e.g. `uses: foo@bar # this part`) as a semver string.
If it can be parsed as such, it will be used to determine if the version
is before the given version. This approach is based on how Dependabot
manages versions for SHA-pinned actions.

To support this change, the workflow/manifest parsing logic had to be
updated to record the comment/annotation. As far as I can tell there's
no way to achieve this with the `yaml:"something"` syntax so I instead
implemented a custom `UnmarshalYAML` function for the `JobStep` struct
which manually parses the `yaml.Node` into a `JobStep` and, if the uses
value is being parsed, stores the line comment into a special field in
the `JobStep` struct.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>