Check reproducibility of the web app build
Add a third reproducibility task that verifies the build for the website
is reproducible. Since `ades.wasm` is the only generated file (others,
such as `COPYING.txt` and `wasm_exec.js`, are just copied, and the rest
is present statically in the repository) it is the only file that is
currently tested.

The reproducibility of the web app is also continuously verified with a
new CI job.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>