Dev: Replace Node.js-based http server by Python-based http server
Remove the dependency on Node.js and npm (as well as the npm package
`http-server`) in favor of using a Python one-liner to use as an HTTP
server for the ades web-app during development. This reduces the number
of direct dependencies, removes the need for a registry dependency
(reducing the relience of Dependabot), and arguably makes contributing
easier as Python seems to be more commonly used across language
ecosystems than Node.js is (though from the provided dev-env experience,
nothing much changes).

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>