Improve the output of violations
This improves upon the output of violations. It does so in a couple of
ways. First it removes the inline suggestions - which become clutter for
many results - in favor of recommending the use of `ades -explain` at
the end of the human-readable report (also resolving the no-op of the
`-suggestions` flag when `-json` is used). Second, it now groups the
violations by the job in which they occur, reducing the amount of
duplicated text in the output.

Additionally this makes the order of the results in the human readable
report deterministic, which is slightly nicer for humans because it is
somewhat sensibly ordered, but notably resolves some flakiness in the
tests.

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>