Filter out expressions that are certainly safe
Update the expressions matcher logic to not complain about expressions
that can be proven to be safe because they consist exclusively of
literals[1] and/or functions[2]  whose output does not contain any of
the input values (i.e. predicate functions as well as `hashFiles`, or
any function operation exclusively on literals).

A fixed point algorithm is used for stripping the safe parts of an
expressions because one iteration might reveal a new safe part. This is
due to the use of multiple regular expression and the fact that these
regular expressions remove only one substring of an expression at the
time. Both of these decisions make the implementation a lot easier to
understand, at the performance cost of the fixed point approach.

The way this is implemented is with a helper function that strips safe
expressions which is used by the actual matchers. The choice to match
twice, once against the input with safe expressions removed and once
against the actual input, may seem unnecessarily bad from a performance
point of view. However, if the stripped input does contain a problem,
we want to get the matches on the original input to present to the user.
The performance impact is minimized by using the `Regexp.Find` method
which will return on the first match.

--
1. https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/evaluate-expressions-in-workflows-and-actions#literals
2. https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/evaluate-expressions-in-workflows-and-actions#functions
--

Signed-off-by: Eric Cornelissen <ericornelissen@gmail.com>